+++ crypto/src/main/java/org/web3j/crypto/exception/CipherException.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.web3j.crypto;
+package org.web3j.crypto.exception;
 
 /** Cipher exception wrapper. */
 public class CipherException extends Exception {
